int type converted to float

+++ assignment2/expressionEvaluation1.py
@@ -2,10 +2,10 @@
 # file name : expressionEvaluation1.py
 # date      : 5/9/20
 
-a = int(input("x = (a - b) / (c - d) \n\nenter a =  "))
-b = int(input("enter b =  "))
-c = int(input("enter c =  "))
-d = int(input("enter d =  "))
+a = float(input("x = (a - b) / (c - d) \n\nenter a =  "))
+b = float(input("enter b =  "))
+c = float(input("enter c =  "))
+d = float(input("enter d =  "))
 
 if(c - d):
     print("x = " + str((a - b) / (c - d)))
